+++ beatmap/v3/parse.ts
@@ -4,12 +4,32 @@ import { Difficulty } from './difficulty.ts';
 import { DifficultyCheck } from './dataCheck.ts';
 import { deepCheck } from '../shared/dataCheck.ts';
 import logger from '../../logger.ts';
+import { IEvent } from '../../types/beatmap/v2/event.ts';
 
 const tag = (name: string) => {
     return `[v3::parse::${name}]`;
 };
 
 const sortObjectTime = (a: IBaseObject, b: IBaseObject) => a.b - b.b;
+
+// temporary fix to CM error
+export function conversionFix(data: Partial<IDifficulty>) {
+    if (data.basicBeatmapEvents) {
+        data.basicBeatmapEvents = data.basicBeatmapEvents.map((ev) => {
+            if ('_time' in ev) {
+                const oldEv = ev as unknown as IEvent;
+                return {
+                    b: oldEv._time,
+                    et: oldEv._type,
+                    i: oldEv._value,
+                    f: oldEv._floatValue ?? 1,
+                    customData: oldEv._customData,
+                };
+            }
+            return ev;
+        });
+    }
+}
 
 export function difficulty(
     data: Partial<IDifficulty>,
@@ -19,24 +39,13 @@ export function difficulty(
     } = { enable: true, throwError: true },
 ): Difficulty {
     logger.info(tag('difficulty'), 'Parsing beatmap difficulty v3.x.x');
-    if (
-        !(
-            data.version === '3.0.0' ||
-            data.version === '3.1.0' ||
-            data.version === '3.2.0'
-        )
-    ) {
+    if (!(data.version === '3.0.0' || data.version === '3.1.0' || data.version === '3.2.0')) {
         logger.warn(tag('difficulty'), 'Unidentified beatmap version');
         data.version = '3.0.0';
     }
+    conversionFix(data);
     if (checkData.enable) {
-        deepCheck(
-            data,
-            DifficultyCheck,
-            'difficulty',
-            data.version,
-            checkData.throwError,
-        );
+        deepCheck(data, DifficultyCheck, 'difficulty', data.version, checkData.throwError);
     }
 
     // haha why do i have to do this, beat games
@@ -52,6 +61,7 @@ export function difficulty(
     data.colorBoostBeatmapEvents = data.colorBoostBeatmapEvents ?? [];
     data.lightColorEventBoxGroups = data.lightColorEventBoxGroups ?? [];
     data.lightRotationEventBoxGroups = data.lightRotationEventBoxGroups ?? [];
+    data.lightTranslationEventBoxGroups = data.lightTranslationEventBoxGroups ?? [];
 
     data.bpmEvents.sort(sortObjectTime);
     data.rotationEvents.sort(sortObjectTime);